Align compilation timing logs with webpack
