+++ crates/rspack_core/src/compilation/build_module_graph/mod.rs
@@ -17,14 +17,10 @@ pub use self::{
   module_executor::{ExecuteModuleId, ExecutedRuntimeModule, ModuleExecutor},
 };
 pub use crate::{BuildModuleGraphArtifact, BuildModuleGraphArtifactState};
-use crate::{Compilation, ExportsInfoArtifact, logger::Logger};
+use crate::{Compilation, ExportsInfoArtifact};
 
 pub async fn build_module_graph_pass(compilation: &mut Compilation) -> Result<()> {
-  let logger = compilation.get_logger("rspack.Compiler");
-  let start = logger.time("build module graph");
-  do_build_module_graph(compilation).await?;
-  logger.time_end(start);
-  Ok(())
+  do_build_module_graph(compilation).await
 }
 
 #[instrument("Compilation:build_module_graph",target=TRACING_BENCH_TARGET, skip_all)]
