+++ crates/rspack_core/src/compilation/finish_make/mod.rs
@@ -1,19 +1,14 @@
 use rspack_error::Result;
 
-use crate::{Compilation, SharedPluginDriver, logger::Logger};
+use crate::{Compilation, SharedPluginDriver};
 
 pub async fn finish_make_pass(
   compilation: &mut Compilation,
   plugin_driver: SharedPluginDriver,
 ) -> Result<()> {
-  let logger = compilation.get_logger("rspack.Compiler");
-  let start = logger.time("finish make hook");
   plugin_driver
     .compiler_hooks
     .finish_make
     .call(compilation)
-    .await?;
-  logger.time_end(start);
-
-  Ok(())
+    .await
 }
